move lesson to WebGoatIntro
