+++ src/it/java/org/owasp/webgoat/integration/WebWolfIntegrationTest.java
@@ -9,6 +9,8 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 import io.restassured.RestAssured;
 import java.util.HashMap;
 import java.util.Map;
+
+import org.apache.commons.lang3.StringUtils;
 import org.junit.jupiter.api.Test;
 
 public class WebWolfIntegrationTest extends IntegrationTest {
@@ -17,23 +19,12 @@ public class WebWolfIntegrationTest extends IntegrationTest {
   public void runTests() {
     startLesson("WebWolfIntroduction");
 
-    // Assignment 3
     Map<String, Object> params = new HashMap<>();
-    params.put("email", this.getUser() + "@webgoat.org");
-      checkAssignment(webGoatUrlConfig.url("WebWolf/mail/send"), params, false);
 
-    String responseBody = readMailbox();
+    String uniqueCode = StringUtils.reverse(this.getUser());
+      params.put("email", this.getUser() + "@webgoat.org");
+        params.put("uniqueCode", uniqueCode);
 
-    String uniqueCode = responseBody.replace("%20", " ");
-    uniqueCode =
-        uniqueCode.substring(
-            21 + uniqueCode.lastIndexOf("your unique code is: "),
-            uniqueCode.lastIndexOf("your unique code is: ") + (21 + this.getUser().length()));
-    params.clear();
-    params.put("uniqueCode", uniqueCode);
-      checkAssignment(webGoatUrlConfig.url("WebWolf/mail"), params, true);
-
-    // Assignment 4
       RestAssured.given()
         .when()
         .relaxedHTTPSValidation()
@@ -50,7 +41,7 @@ public class WebWolfIntegrationTest extends IntegrationTest {
         .get(webWolfUrlConfig.url("landing"))
         .then()
         .statusCode(200);
-    responseBody =
+    String responseBody =
         RestAssured.given()
             .when()
             .relaxedHTTPSValidation()
