+++ src/main/java/org/owasp/webgoat/lessons/webgoatintroduction/Email.java
@@ -2,7 +2,7 @@
  * SPDX-FileCopyrightText: Copyright © 2017 WebGoat authors
  * SPDX-License-Identifier: GPL-2.0-or-later
  */
-package org.owasp.webgoat.lessons.webwolfintroduction;
+package org.owasp.webgoat.lessons.webgoatintroduction;
 
 import java.io.Serializable;
 import lombok.Builder;
