+++ src/main/resources/i18n/messages_nl.properties
@@ -47,6 +47,6 @@ register.condition.1=Wanneer u WebGoat runt op uw computer, bent u kwetsbaar voo
   Zorg dat u geen verbinding heeft met internet en dat toegang tot WebGoat alleen lokaal mogelijk is om het aanvalsoppervlak te verkleinen.
 register.condition.2=WebGoat is bedoeld als educatieve applicatie op het gebied van secure software development. \
   Gebruik wat u leert om applicaties beter te maken en niet om zonder toestemming applicaties te schaden. \
-  In dat laatste geval loopt u risico op rechtsvervoling en ontslag.
+  In dat laatste geval loopt u risico op rechtsvervolging en ontslag.
 
 mailbox=Postvak
