+++ src/main/resources/lessons/challenges/html/Challenge7.html
@@ -43,7 +43,7 @@ f94008f801fceb8833a30fe56a8b26976347edcf First version of WebGoat Cloud website
                                                    value="Reset Password" type="submit"/>
                                         </div>
                                         <div class="form-group">
-                                            <p>(c) 2023 WebGoat Cloud Platform</p>
+                                            <p>(c) 2026 WebGoat Cloud Platform</p>
                                         </div>
 
                                         <input type="hidden" class="hide" name="token" id="token" value=""/>
