+++ src/main/resources/lessons/webgoatintroduction/documentation/Receiving_mail.adoc
@@ -5,7 +5,7 @@ You can open it with the envelope (mailbox) button in the top right corner,
 next to the language selector. The mailbox opens in a separate window and the button
 shows a badge with the number of unread e-mails.
 This mailbox is user-specific, so each user has a separate mailbox. All e-mail
-sent to {user}@.... will end up in this inbox.
+sent to `username:user[]`@webgoat.org will end up in this inbox.
 
 {nbsp}
 {nbsp}
