+++ src/main/resources/lessons/webgoatintroduction/i18n/WebGoatLabels.properties
@@ -1,1 +1,5 @@
 webgoat.title=WebGoat
+webgoat.code_incorrect=That is not the correct code: {0}, please try again.
+webgoat.email_failed=There was an error while sending the e-mail. Is WebWolf running?
+webgoat.email_send=An email has been send to {0} please check your inbox.
+webgoat.email_mismatch=Of course you can send mail to user {0} however you will not be able to read this e-mail in WebWolf, please use your own username.
