+++ src/main/resources/lessons/webgoatintroduction/i18n/WebGoatLabels_nl.properties
@@ -1,1 +1,5 @@
 webgoat.title=WebGoat
+webgoat.code_incorrect=Dit {0} was niet de juiste code, probeer het opnieuw.
+webgoat.email_failed=Het versturen van de mail is mislukt.
+webgoat.email_send=Er is een email verstuurd naar {0} check je inbox.
+webgoat.email_mismatch=Natuurlijk mag je proberen mail te sturen naar gebruiker {0} maar dat kun je dan niet bekijken, gebruik je eigen usernaam@webgoat.org.
