+++ src/main/resources/lessons/webgoatintroduction/images/menuheader.png
