+++ src/main/resources/lessons/webwolfintroduction/i18n/WebGoatLabels.properties
@@ -1,9 +1,4 @@
 webwolf.title=WebWolf
 
-webwolf.email_send=An email has been send to {0} please check your inbox.
-webwolf.code_incorrect=That is not the correct code: {0}, please try again.
-webwolf.email_failed=There was an error while sending the e-mail. Is WebWolf running?
-
-webwolf.email_mismatch=Of course you can send mail to user {0} however you will not be able to read this e-mail in WebWolf, please use your own username.
 
 webwolf.landing_wrong=This is the wrong code, try to look for the uniqueCode in the parameters in WebWolf.
