+++ src/main/resources/lessons/webwolfintroduction/templates/webwolfPasswordReset.html
@@ -23,7 +23,7 @@
                     </div>
                 </div>
                 <div>
-                    <a href="https://github.com/WebGoat">(c) 2023 WebGoat Company</a>
+                    <a href="https://github.com/WebGoat">(c) 2026 WebGoat Company</a>
                 </div>
             </form>
         </div>
