+++ src/main/resources/webwolf/templates/fragments/footer.html
@@ -7,7 +7,7 @@
 
     <div class="container">
         <footer>
-            © 2017 - 2023 WebGoat - Use WebWolf at your own risk
+            © 2017 - 2026 WebGoat - Use WebWolf at your own risk
         </footer>
     </div>
